requirements.txt
asgiref==3.3.4
Django==3.2.3
django-rest-passwordreset==1.1.0
djangorestframework==3.12.4
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
